add field created by

+++ src/main/java/home/test/controller/DocumentController.java
@@ -62,14 +62,26 @@ public class DocumentController {
 
     @RequestMapping(value = "/editFolder", method = RequestMethod.GET)
     public ModelAndView showEditFolderForm(@RequestParam(required = true) Long id){
-        return new ModelAndView("addFolder", "folder", folderService.get(id));
+        Folder folder = folderService.get(id);
+        if (folder.getCreatedBy().equals(SecurityContextHolder.getContext().getAuthentication().getName())){
+            return new ModelAndView("addFolder", "folder", folder);
+        }
+        else
+            return new ModelAndView("");
     }
 
     @RequestMapping(value = "/deleteFolder", method = RequestMethod.GET)
     public String deleteFolder(@RequestParam(required = true) Long id){
-        Long parentId = folderService.get(id).getParentId();
-        folderService.remove(id);
+        Folder folder = folderService.get(id);
+        Long parentId = folder.getParentId();
+        if (folder.getCreatedBy().equals(SecurityContextHolder.getContext().getAuthentication().getName())){
+            folderService.remove(id);
+        }
+        else{
+
+        }
         return "redirect:/document?id=" + parentId;
+
     }
 
     @RequestMapping(value = "/deleteDoc", method = RequestMethod.GET)
